+++ packages/carbon-graphs/dev/examples/popup.js
@@ -81,6 +81,9 @@ const renderY2Popup = (fn) => {
 const removeOldPopup = () => {
     // Remove old popup
     d3.select("#overlay").remove();
+    if(document.getElementsByClassName('initial-tooltip')[0]) {
+        document.getElementsByClassName('initial-tooltip')[0].className = 'tooltip';
+    }
     d3.select("#tooltip").attr("style", "").selectAll("g").remove();
 };
 export const loadPopup = (onCloseCB, key, index, value) => {