+++ packages/terra-graphs-docs/src/example-datasets/dataObjects/General/noDataView.js
@@ -1,0 +1,18 @@
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import {
+  loadPopup,
+} from '@cerner/terra-graphs-docs/lib/example-datasets/popup';
+
+const data = [
+  {
+    key: 'uid_1',
+    label: {
+      display: 'Data Label 1',
+    },
+    color: Carbon.helpers.COLORS.BLACK,
+    onClick: loadPopup,
+    values: [],
+  },
+];
+
+export default data;