+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/General/Legend/LineWithLegendOption.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
+import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib//components/Dev.module.scss';
+import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/defaultMultiLineData';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(lineDefault('#lineWithLegendOption'));
+
+const contentData = utils.deepClone(data);
+contentData[0].showShapes = false;
+contentData[0].color = Carbon.helpers.COLORS.BLUE;
+contentData[0].style = { strokeDashArray: '2,2' };
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <LineGraph graphID="lineWithLegendOption" graphConfig={graphConfig} dataset={contentData} />
+  </React.Fragment>
+);
+