+++ packages/terra-graphs-docs/src/terra-dev-site/doc/example/Graphs/LineGraph/MultiLine/Simple.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib/components/Dev.module.scss';
+import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/multiLineData';
+
+/*
+Please refer to the documentation below to see the graphConfig and data objects
+*/
+const graphConfig = utils.deepClone(lineDefault('#multiLine'));
+const dataset = utils.deepClone(data);
+const timeout = [0, 750, 750 * 2, 750 * 3, 750 * 4, 750 * 5, 750 * 6];
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <LineGraph graphID="multiLine" graphConfig={graphConfig} dataset={dataset} timeout={timeout} />
+  </React.Fragment>
+);