+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/General/Axes/SuppressTickValueTrailingZero.test.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib/components/Dev.module.scss';
+import lineDecimalAxes from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/General/lineDecimalAxesConfig';
+
+const graphConfig = utils.deepClone(lineDecimalAxes('#supressTickValue'));
+
+export default () => <LineGraph graphID="supressTickValue" graphConfig={graphConfig} />;