+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/LineGraph/Regions/Line/NoLowerBound.test.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import LineGraph from '@cerner/terra-graphs/lib/components/Line/LineGraph';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib/components/Dev.module.scss';
+import lineDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Line/lineDefault';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Line/simpleLineData';
+
+const graphConfig = utils.deepClone(lineDefault('#noLowerBound'));
+const contentData = utils.deepClone(data);
+contentData[0].regions = [
+  {
+    end: 10,
+  },
+];
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <LineGraph graphID="noLowerBound" graphConfig={graphConfig} dataset={contentData} />
+  </React.Fragment>
+);