+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/PairedResult/Regions/Simple.test.jsx
@@ -1,0 +1,41 @@
+import React from 'react';
+import PairedResultGraph from '@cerner/terra-graphs/lib/components/PairedResult/PairedResultGraph';
+import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
+import '@cerner/terra-graphs/lib/components/Dev.module.scss';
+import pairedResultDefault from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/PairedResult/pairedResultDefault';
+import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/PairedResult/pairedResultSimpleData';
+
+const regions = {
+  high: [
+    {
+      axis: 'y',
+      start: 140,
+      end: 220,
+      color: '#c8cacb',
+    },
+  ],
+  low: [
+    {
+      axis: 'y',
+      start: 20,
+      end: 70,
+    },
+  ],
+};
+/*
+Please refer documentation below to see graphConfig and data objects.
+*/
+const graphConfig = utils.deepClone(pairedResultDefault('#simplePairedResultGraph'));
+const contentData = utils.deepClone(data);
+
+contentData[0].regions = {
+  high: [regions.high[0]],
+  low: regions.low,
+};
+
+export default () => (
+  <React.Fragment>
+    <div id="tooltip" className="initial-tooltip" />
+    <PairedResultGraph graphID="simplePairedResultGraph" graphConfig={graphConfig} dataset={contentData} />
+  </React.Fragment>
+);