+++ packages/terra-graphs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Create Components, examples and WDIO test cases for Line, Bar and Paired Result graph types.
+  
 * Added
   * Added documentation for `reflowMultipleDatasets`.